Slash command compile 1.0
