+++ src/handler.ts
@@ -1,6 +1,5 @@
 import { ActionObject } from "./types";
 import { SlashCommandBuilder } from "@discordjs/builders";
-import { report } from "./utility";
 import { Router } from "./router";
 
 type HandlerContent = ActionObject | Router;
@@ -14,7 +13,7 @@ type RegexActions = Map<RegExp, HandlerContent>;
 type TriggerType = string | RegExp;
 
 export class CommandsHandler {
-  private stringActions: MessageActions;
+  readonly stringActions: MessageActions;
   private regexActions: RegexActions;
   private defaultAction: ActionObject | undefined;
   readonly commands: ReturnType<SlashCommandBuilder["toJSON"]>[];
@@ -27,122 +26,6 @@ export class CommandsHandler {
     this.findAction = this.findAction.bind(this);
     this.setAction = this.setAction.bind(this);
     this.removeAction = this.removeAction.bind(this);
-    this.createSlashCommandParams = this.createSlashCommandParams.bind(this);
-  }
-
-  //Creates a new Slash command JSON to send to the API
-  createSlashCommandParams(
-    name: string,
-    description: string,
-    parameters: ActionObject["parameters"]
-  ) {
-    const isUppercase = (t: string) => /[A-Z]/.test(t);
-    if (isUppercase(name))
-      throw new Error(
-        `[CreateSlashCommandParams] => Trigger "${name}" must not have uppercase letters`
-      );
-
-    const command = new SlashCommandBuilder()
-      .setName(name)
-      .setDescription(description || "A command");
-
-    if (parameters) {
-      parameters.forEach((param) => {
-        if (isUppercase(param.name))
-          throw new Error(
-            `[CreateSlashCommandParams] => Parameter "${param.name}" must not have uppercase letters`
-          );
-        switch (param.type ?? "STRING") {
-          case "STRING":
-            {
-              command.addStringOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-          case "BOOLEAN":
-            {
-              command.addBooleanOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-          case "INTEGER":
-            {
-              command.addIntegerOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-          case "MENTIONABLE":
-            {
-              command.addMentionableOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-
-          case "NUMBER":
-            {
-              command.addNumberOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-
-          case "ROLE":
-            {
-              command.addRoleOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-
-          case "USER":
-            {
-              command.addUserOption((option) =>
-                option
-                  .setName(param.name)
-                  .setDescription(description || "A command")
-                  .setRequired(true)
-              );
-            }
-            break;
-
-          default:
-            report(
-              `[CommandsHandler] => Parameter type "${param.type}" is unsupported`
-            );
-        }
-      });
-    } else {
-      command.addStringOption((option) =>
-        option
-          .setName("arguments")
-          .setRequired(true)
-          .setDescription(description || "A command")
-      );
-    }
-
-    return command.toJSON();
   }
 
   setAction(trigger: TriggerType, action: HandlerContent) {
